chore(parser): move LimitTracker to its own module
